add solution for 4.21*: bloglist expansion, step9

+++ part4/blog/controllers/blogs.js
@@ -42,6 +42,25 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+	const decodedToken = jwt.verify(request.token, process.env.SECRET)
+	if (!decodedToken.id) {
+		return response.status(401).json({ error: 'Invalid or missing token' })
+	}
+  
+	const user = await User.findById(decodedToken.id)
+	if (!user) {
+		return response.status(404).json({ error: 'User not found' })
+	}
+
+	const blog = await Blog.findById(request.params.id)
+	if (!blog) {
+		return response.status(404).json({ error: 'Blog not found' })
+	}
+
+	if (blog.user.toString() !== user.id.toString()) {
+		return response.status(403).json({ error: 'Unauthorized to delete this blog' })
+	}
+
 	await Blog.findByIdAndDelete(request.params.id)
 	response.status(204).end()
 })
